move program into a folder
